# controller/adminController.js
import prisma from "../lib/prisma.js";
import jwt from "jsonwebtoken";
import { calculatePriority, getMaxRepostCount } from "../utils/priority.js";

export class AdminController {
  // now we need to authenticate the admin using the email and password itself
  static login = async (req, res) => {
    try {
    console.log("this has been started" , req.body)
      const { email, password } = req.body;
      const admin = await prisma.admin.findUnique({ where: { email : email } });
      console.log("admin is this" , admin)
      if (!admin) {
        return res.status(401).json({ message: "Admin not found" });
      }
      if (admin.password !== password) {
        return res.status(401).json({ message: "Invalid password" });
      }

      if (!process.env.JWT_SECRET) {
        console.error("[Admin] JWT_SECRET missing");
        return res.status(500).json({ message: "Server misconfiguration." });
      }

      // Generate JWT token with admin role
      const token = jwt.sign(
        {
          sub: admin.id,
          email: admin.email,
          role: "admin",
        },
        process.env.JWT_SECRET,
        { expiresIn: process.env.JWT_EXPIRES_IN ?? "12h" }
      );
      
      return res.status(200).json({
        success: true,
        message: "Admin logged in successfully",
        token, // JWT token with admin role
        admin: {
          id: admin.id,
          email: admin.email,
          role: "admin",
        },
      });
    } catch (error) {
      console.error("[Admin] login failed:", error);
      return res.status(500).json({ message: "Login failed" , error : error});
    }
  };

  // Update issue status (0-5 scale)
  static updateIssueStatus = async (req, res) => {
    try {
      const { id } = req.params;
      const { status } = req.body;

      // Validate status is a number between 0-5
      const statusValue = parseInt(status);
      if (isNaN(statusValue) || statusValue < 0 || statusValue > 5) {
        return res.status(400).json({
          message: "Status must be a number between 0 and 5 (0=pending, 5=resolved)",
        });
      }
      
      const issue = await prisma.issue.findUnique({
        where: { id },
      });

      if (!issue) {
        return res.status(404).json({ message: "Issue not found" });
      }

      const updatedIssue = await prisma.issue.update({
        where: { id },
        data: { status: statusValue },
        include: {
          category: true,
          user: {
            select: {
              id: true,
              number: true,
            },
          },
          comments: {
            include: {
              admin: {
                select: {
                  id: true,
                  email: true,
                },
              },
            },
            orderBy: {
              createdAt: "asc",
            },
          },
          reposts: {
            select: {
              id: true,
              userId: true,
            },
          },
        },
      });

      // Calculate priority
      const maxReposts = await getMaxRepostCount(prisma);
      const repostCount = updatedIssue.reposts.length;
      const priority = calculatePriority(updatedIssue.importanceRating, repostCount, maxReposts);

      return res.status(200).json({
        message: "Issue status updated successfully",
        issue: {
          ...updatedIssue,
          repostCount,
          priority,
        },
      });
    } catch (error) {
      console.error("[Admin] updateIssueStatus failed:", error);
      return res.status(500).json({ message: "Failed to update issue status" });
    }
  };

  // Get all issues for admin dashboard
  static getIssues = async (req, res) => {
    try {
      const { status, categoryId, sortBy = "createdAt", order = "desc" } = req.query;

      const where = {};
      if (status !== undefined) {
        // allow numeric (0-5) or enum string
        if (!Number.isNaN(parseInt(status))) {
          const s = parseInt(status);
          if (s < 0 || s > 5) {
            return res.status(400).json({ message: "Status must be between 0 and 5" });
          }
          where.status = s === 0 ? "PENDING" : s === 5 ? "COMPLETED" : "PROGRESS";
        } else {
          where.status = status.toUpperCase();
        }
      }
      if (categoryId) {
        where.categoryId = categoryId;
      }

      const issues = await prisma.issue.findMany({
        where,
        include: {
          category: true,
          user: {
            select: {
              id: true,
              number: true,
            },
          },
          comments: {
            include: {
              admin: {
                select: {
                  id: true,
                  email: true,
                },
              },
            },
            orderBy: {
              createdAt: "asc",
            },
          },
          reposts: {
            select: {
              id: true,
              userId: true,
            },
          },
        },
        orderBy: {
          [sortBy]: order,
        },
      });

      // Get max repost count for normalization
      const maxReposts = await getMaxRepostCount(prisma);

      // Calculate priority for each issue
      const issuesWithPriority = issues.map((issue) => {
        const repostCount = issue.reposts.length;
        const priority = calculatePriority(issue.importanceRating, repostCount, maxReposts);
        
        return {
          ...issue,
          repostCount,
          priority,
        };
      });

      // Sort by priority if requested
      if (sortBy === "priority") {
        issuesWithPriority.sort((a, b) => {
          return order === "desc"
            ? b.priority.totalPriority - a.priority.totalPriority
            : a.priority.totalPriority - b.priority.totalPriority;
        });
      }

      return res.status(200).json({
        message: "Issues retrieved successfully",
        issues: issuesWithPriority,
        count: issuesWithPriority.length,
      });
    } catch (error) {
      console.error("[Admin] getIssues failed:", error);
      return res.status(500).json({ message: "Failed to retrieve issues" });
    }
  };

  // Get single issue by ID for admin
  static getIssueById = async (req, res) => {
    try {
      const { id } = req.params;

      const issue = await prisma.issue.findUnique({
        where: { id },
        include: {
          category: true,
          user: {
            select: {
              id: true,
              number: true,
            },
          },
          comments: {
            include: {
              admin: {
                select: {
                  id: true,
                  email: true,
                },
              },
            },
            orderBy: {
              createdAt: "asc",
            },
          },
          reposts: {
            select: {
              id: true,
              userId: true,
            },
          },
        },
      });

      if (!issue) {
        return res.status(404).json({ message: "Issue not found" });
      }

      // Calculate priority
      const maxReposts = await getMaxRepostCount(prisma);
      const repostCount = issue.reposts.length;
      const priority = calculatePriority(issue.importanceRating, repostCount, maxReposts);

      return res.status(200).json({
        message: "Issue retrieved successfully",
        issue: {
          ...issue,
          repostCount,
          priority,
        },
      });
    } catch (error) {
      console.error("[Admin] getIssueById failed:", error);
      return res.status(500).json({ message: "Failed to retrieve issue" });
    }
  };
}   