# openapi.yaml
openapi: 3.0.0
info:
  title: Hack Backend API
  description: |
    API for civic issue reporting and management system.
    - Users can report issues, view their issues, and repost issues
    - Admins can manage issues, update status, and add comments
    - Authentication via JWT tokens (Bearer token)
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
...
  /user/issues/{id}:
    get:
      tags:
        - User Issues
      summary: Get issue by ID
      description: Retrieve a specific issue by ID
      operationId: getIssueById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Issue ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Issue retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  issue:
                    $ref: '#/components/schemas/Issue'
        '401':
          description: Unauthorized
        '404':
          description: Issue not found
        '500':
          description: Server error

  # Admin Issue Endpoints
...
  /admin/issues/{id}/status:
    put:
      tags:
        - Admin Issues
      summary: Update issue status
      description: Update the status of an issue (0=pending, 5=resolved)
      operationId: updateIssueStatus
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Issue ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: integer
                  minimum: 0
                  maximum: 5
                  description: Status value (0=pending, 5=resolved)
                  example: 3
      responses:
        '200':
          description: Issue status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  issue:
                    $ref: '#/components/schemas/Issue'
        '400':
          description: Invalid status value
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin access required
        '404':
          description: Issue not found
        '500':
          description: Server error

  # Comment Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoints
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Infrastructure"
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Issue:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        image:
          type: string
          format: uri
          nullable: true
        importanceRating:
          type: integer
          minimum: 0
          maximum: 5
          description: User's importance rating (0-5)
        status:
          type: integer
          minimum: 0
          maximum: 5
          description: Admin status (0=pending, 5=resolved)
        categoryId:
          type: string
          format: uuid
        category:
          $ref: '#/components/schemas/Category'
        userId:
          type: string
          format: uuid
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            number:
              type: string
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        reposts:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              userId:
                type: string
                format: uuid
        repostCount:
          type: integer
          description: Number of reposts
        priority:
          type: object
          properties:
            totalPriority:
              type: number
              description: Total priority score (0-10)
            normalizedPriority:
              type: number
              description: Normalized priority (0-5 scale)
            userRatingPoints:
              type: number
            repostPoints:
              type: number
            repostCount:
              type: integer
            importanceRating:
              type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        issueId:
          type: string
          format: uuid
        issue:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            title:
              type: string
        adminId:
          type: string
          format: uuid
          nullable: true
          description: Present if comment is from an admin
        admin:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
        userId:
          type: string
          format: uuid
          nullable: true
          description: Present if comment is from a user
        user:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            number:
              type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...